Improve on @donjohnson's ideas
First, we look at the system installed types.db (works on centos
package, should probably be generalized).

Later, we split up the metric name into is parts according to the
collectd data spec. The instance are split out and handled somewhat
better.

This fixes an issue where type instances will break the lookup in
the types.db file.

This also fixes an issue where metrics will be poorly named when
they appear in graphite.

CAVEAT: This makes changes that are not compatible with existing
proxy users. The graphs will be given different names.

tl:dr This breaks backwards compatability